pyproject.toml
[project]
name = "Badger"
version = "0.1.0"
description = "A system to print various kinds of Badges and labels."
authors = [
    {name = "NeoCortex97", email = "stefan_schmelz@gmx.net"},
]
dependencies = [
    "fastapi~=0.115.6",
    "spacedirectory~=0.0.5",
    "pillow~=11.0.0",
    "brother-ql",
    "qrcode[pil]~=8.0",
    "python-escpos",
    "sqlmodel~=0.0.22",
    "starlette>=0.40.0",
    "pydantic~=2.10.3",
    "pyzmq~=26.2.0",
    "imutils~=0.5.4",
    "matplotlib~=3.10.0",
    "numpy~=2.2.0",
    "opencv-python~=4.10.0.84",
    "requests~=2.32.3",
    "urllib3~=2.2.3",
    "rich~=13.9.4",
    "python-dotenv~=1.0.1",
    "appdirs~=1.4.4",
    "annotated-types~=0.7.0",
    "netifaces>=0.11.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
